requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pypdf>=4.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

